refactor: remove openAuthModal and closeAuthModal methods from auth context
