+++ src/containers/Layout/ConnectButton/index.tsx
@@ -1,5 +1,6 @@
 import { Button, ButtonProps } from 'components';
 import { useTranslation } from 'packages/translations';
+import { useAuthModal } from 'packages/wallet';
 import { truncateAddress } from 'utilities';
 
 import { useGetPrimeToken } from 'clients/api';
@@ -38,7 +39,8 @@ export const ConnectButtonUi: React.FC<ConnectButtonUiProps> = ({
 export const ConnectButton: React.FC<
   Omit<ConnectButtonUiProps, 'isAccountPrime' | 'accountAddress' | 'loading'>
 > = props => {
-  const { accountAddress, openAuthModal } = useAuth();
+  const { accountAddress } = useAuth();
+  const { openAuthModal } = useAuthModal();
 
   const { data: getPrimeTokenData, isLoading: isGetPrimeTokenLoading } = useGetPrimeToken({
     accountAddress,
