+++ src/context/AuthContext/index.tsx
@@ -5,16 +5,12 @@ import React, { useContext } from 'react';
 import { ChainId } from 'types';
 
 export interface AuthContextValue {
-  openAuthModal: () => void;
-  closeAuthModal: () => void;
   switchChain: (input: { chainId: ChainId }) => void;
   chainId: ChainId;
   accountAddress?: string;
 }
 
 export const AuthContext = React.createContext<AuthContextValue>({
-  openAuthModal: noop,
-  closeAuthModal: noop,
   switchChain: noop,
   chainId: store.getState().chainId,
 });
@@ -30,16 +26,10 @@ export const AuthProvider: React.FC<AuthProviderProps> = ({ children }) => {
   const { chainId } = useChainId();
   const { accountAddress } = useAccountAddress();
 
-  const setIsAuthModalOpen = store.use.setIsAuthModalOpen();
-  const closeAuthModal = () => setIsAuthModalOpen({ isAuthModalOpen: false });
-  const openAuthModal = () => setIsAuthModalOpen({ isAuthModalOpen: true });
-
   return (
     <AuthContext.Provider
       value={{
         accountAddress,
-        openAuthModal,
-        closeAuthModal,
         switchChain,
         chainId,
       }}
