+++ src/packages/wallet/hooks/useAuthModal/__mocks__/index.tsx
@@ -1,0 +1,5 @@
+export const useAuthModal = vi.fn(() => ({
+  isAuthModalOpen: false,
+  openAuthModal: vi.fn(),
+  closeAuthModal: vi.fn(),
+}));
