+++ src/packages/wallet/hooks/useProvider/__mocks__/index.ts
@@ -1,4 +1,6 @@
 import { providers } from 'ethers';
 
 const fakeProvider = new providers.JsonRpcProvider();
-export const useProvider = vi.fn(() => fakeProvider);
+export const useProvider = vi.fn(() => ({
+  provider: fakeProvider,
+}));
