+++ src/packages/wallet/hooks/useSigner/__mocks__/index.ts
@@ -1,3 +1,3 @@
-import fakeSigner from '__mocks__/models/signer';
-
-export const useSigner = vi.fn(() => fakeSigner);
+export const useSigner = vi.fn(() => ({
+  signer: undefined,
+}));
