+++ src/pages/Dashboard/Banner/__tests__/index.spec.tsx
@@ -1,23 +1,16 @@
 import { fireEvent } from '@testing-library/react';
 import { en } from 'packages/translations';
+import { useAuthModal } from 'packages/wallet';
 import Vi from 'vitest';
 
 import fakeAccountAddress from '__mocks__/models/address';
 import { useGetPrimeToken } from 'clients/api';
-import { useAuth } from 'context/AuthContext';
 import { renderComponent } from 'testUtils/render';
 
 import { Banner } from '..';
 
-vi.mock('context/AuthContext');
-
 describe('Banner', () => {
   beforeEach(() => {
-    (useAuth as Vi.Mock).mockImplementation(() => ({
-      accountAddress: undefined,
-      openAuthModal: vi.fn(),
-    }));
-
     (useGetPrimeToken as Vi.Mock).mockImplementation(() => ({
       data: {
         exists: false,
@@ -37,11 +30,11 @@ describe('Banner', () => {
   });
 
   it('displays nothing when user is connected', () => {
-    (useAuth as Vi.Mock).mockImplementation(() => ({
-      accountAddress: fakeAccountAddress,
-    }));
-
-    const { baseElement } = renderComponent(<Banner />);
+    const { baseElement } = renderComponent(<Banner />, {
+      authContextValue: {
+        accountAddress: fakeAccountAddress,
+      },
+    });
 
     expect(baseElement.textContent).toEqual('');
   });
@@ -49,8 +42,8 @@ describe('Banner', () => {
   it('opens auth modal when clicking on button', () => {
     const openAuthModalMock = vi.fn();
 
-    (useAuth as Vi.Mock).mockImplementation(() => ({
-      accountAddress: undefined,
+    (useAuthModal as Vi.Mock).mockImplementation(() => ({
+      isAuthModalOpen: false,
       openAuthModal: openAuthModalMock,
     }));
