+++ src/pages/Governance/VotingWallet/index.tsx
@@ -12,6 +12,7 @@ import {
 } from 'components';
 import { useGetToken } from 'packages/tokens';
 import { useTranslation } from 'packages/translations';
+import { useAuthModal } from 'packages/wallet';
 import React, { useMemo, useState } from 'react';
 import { Token } from 'types';
 import { areTokensEqual, convertMantissaToTokens } from 'utilities';
@@ -215,7 +216,8 @@ export const VotingWalletUi: React.FC<VotingWalletUiProps> = ({
 
 const VotingWallet: React.FC = () => {
   const [delegateModelIsOpen, setDelegateModelIsOpen] = useState(false);
-  const { accountAddress, openAuthModal } = useAuth();
+  const { accountAddress } = useAuth();
+  const { openAuthModal } = useAuthModal();
   const xvs = useGetToken({
     symbol: 'XVS',
   });
